+++ delivery/pipelines/run.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 function run() {
-  docker-compose run --rm app $*
+  docker-compose run --rm app "$@"
 }
 
-source "$(dirname "$0")/../base.inc"
+run "$@"