+++ test/test_cli.py
@@ -6,8 +6,16 @@ from app.cli import cli
 
 
 @mock.patch("app.cli.main")
-def test_main_ok(main_m):
+def test_main_normal_ok(main_m):
     result = CliRunner().invoke(cli)
     assert result.exit_code == 0
 
     main_m.assert_called_once_with()
+
+
+@mock.patch("app.cli.main")
+def test_main_update_single_source_ok(main_m):
+    result = CliRunner().invoke(cli, ["update-single-source", "test"])
+    assert result.exit_code == 0
+
+    main_m.assert_called_once_with("test")